ai/backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
gTTS
SpeechRecognition
python-dotenv
azure-search-documents
openai
